requirements.txt
# Core dependencies
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0

# Speech processing
faster-whisper==0.10.0  # More efficient than openai-whisper
speechrecognition==3.10.0

# OpenAI
openai==1.12.0  # Updated from 1.0.0

# Additional utilities
websockets==12.0  # Required for WebSocket support
pydantic==2.6.1  # For data validation
numpy==1.26.3  # Required for faster-whisper